+++ Modules/Purchase/Database/Migrations/2021_08_08_021108_create_purchases_table.php
@@ -22,11 +22,11 @@ class CreatePurchasesTable extends Migration
             $table->integer('tax_percentage')->default(0);
             $table->integer('tax_amount')->default(0);
             $table->integer('discount_percentage')->default(0);
-            $table->integer('discount_amount')->default(0);
-            $table->integer('shipping_amount')->default(0);
-            $table->integer('total_amount');
-            $table->integer('paid_amount');
-            $table->integer('due_amount');
+            $table->unsignedBigInteger('discount_amount')->default(0);
+            $table->unsignedBigInteger('shipping_amount')->default(0);
+            $table->unsignedBigInteger('total_amount');
+            $table->unsignedBigInteger('paid_amount');
+            $table->bigInteger('due_amount');
             $table->string('status');
             $table->string('payment_status');
             $table->string('payment_method');
