+++ Modules/PurchasesReturn/Database/Migrations/2021_08_08_222646_create_purchase_return_payments_table.php
@@ -16,7 +16,7 @@ class CreatePurchaseReturnPaymentsTable extends Migration
         Schema::create('purchase_return_payments', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('purchase_return_id');
-            $table->integer('amount');
+            $table->unsignedBigInteger('amount');
             $table->date('date');
             $table->string('reference');
             $table->string('payment_method');
