+++ Modules/Sale/Database/Migrations/2021_07_31_201003_create_sales_table.php
@@ -21,13 +21,13 @@ class CreateSalesTable extends Migration
             $table->unsignedBigInteger('customer_id')->nullable();
             $table->string('customer_name');
             $table->integer('tax_percentage')->default(0);
-            $table->integer('tax_amount')->default(0);
+            $table->unsignedBigInteger('tax_amount')->default(0);
             $table->integer('discount_percentage')->default(0);
             $table->unsignedBigInteger('discount_amount')->default(0);
             $table->unsignedBigInteger('shipping_amount')->default(0);
             $table->unsignedBigInteger('total_amount');
             $table->unsignedBigInteger('paid_amount');
-            $table->integer('due_amount');
+            $table->bigInteger('due_amount');
             $table->string('status');
             $table->string('payment_status');
             $table->string('payment_method');
